openai_assistants_api_demo/pyproject.toml
[tool.poetry]
name = "openai-assistants-api-demo"
version = "0.1.0"
description = ""
authors = ["kirin <kirin.chen1001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
